Seperated code into platform layers
